# Tab logic cleanup