--- a/mittab/apps/tab/team_views.py
+++ b/mittab/apps/tab/team_views.py
@@ -340,24 +340,12 @@ def rank_teams_ajax(request):
                               context_instance=RequestContext(request))
 
 def rank_teams(request):
-    print("starting rankings: ", datetime.now())
-    ranked_teams = tab_logic.rank_teams()
-    print("Got ranked teams")
-    teams = [(team,
-              tab_logic.tot_wins(team),
-              tab_logic.tot_speaks(team),
-              tab_logic.tot_ranks(team))
-              for team in ranked_teams]
+    ranked_teams = tab_logic.rankings.rank_teams()
+    teams = [(ts.team, ts.wins, ts.speaks, ts.ranks) for ts in ranked_teams]
 
-    print("started novice rankings: ", datetime.now())
-    ranked_novice_teams = tab_logic.rank_nov_teams()
-    nov_teams = [(team,
-                  tab_logic.tot_wins(team),
-                  tab_logic.tot_speaks(team),
-                  tab_logic.tot_ranks(team))
-                  for team in ranked_novice_teams]
-
-    print("Got ranked novice teams")
+    nov_teams = filter(
+            lambda ts: all(map(lambda d: d.novice_status == Debater.NOVICE, ts[0].debaters.all())),
+            teams)
     return render_to_response('rank_teams_component.html',
                              {'varsity': teams,
                               'novice': nov_teams,