… mittab/libs/tab_logic/stats.py
… mittab/libs/tests/test_tab_logic.py