This PR cleans up the tab logic library into 1 main library with some sub-modules. Goals are to separate:

- stats calculations
- rankings
- pairing

And also to improve performance