@BenMusch 2019-03-26
Some performance stats:

Loading speaker rankings went from ~16s to ~6s

Loading debater rankings went from ~17s to ~7s

Curious @ifly6: Is that about comparable to what you observed?