@ifly6 2019-03-26
I was able to halve runtime by introducing struts. There's a number of Django query optimisations I also introduced in my pull request.